% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineConceptSetsInCohortDefinitionSet.R
\name{combineConceptSetsInCohortDefinitionSet}
\alias{combineConceptSetsInCohortDefinitionSet}
\title{combine concept sets}
\usage{
combineConceptSetsInCohortDefinitionSet(cohortDefinitionSet)
}
\value{
Returns a tibble data frame.
}
\description{
given a cohort definition set, this function extracts the concept set json and sql for all cohorts,
compares concept sets across cohort definitions, assigns unique id.
}
