% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSetDiagnostics.R
\docType{package}
\name{ConceptSetDiagnostics-package}
\alias{ConceptSetDiagnostics}
\alias{ConceptSetDiagnostics-package}
\title{ConceptSetDiagnostics: Concept Set Diagnostics}
\description{
A package to work with OMOP concepts and cohort concept sets.
}
\keyword{internal}
