% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertConceptSetExpressionToDataFrame.R
\name{convertConceptSetExpressionToDataFrame}
\alias{convertConceptSetExpressionToDataFrame}
\title{convert a concept set expression object into a data frame object}
\usage{
convertConceptSetExpressionToDataFrame(
  conceptSetExpression,
  updateVocabularyFields = FALSE,
  connection = NULL,
  connectionDetails = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  vocabularyDatabaseSchema = "vocabulary"
)
}
\arguments{
\item{conceptSetExpression}{An R-object (list) with expression of the concept set.}

\item{updateVocabularyFields}{Do you want to update the details of concepts from the vocabulary tables? If yes,
then connection or connectionDetails to a remote db with OMOP vocabulary tables
is needed.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}

\item{vocabularyDatabaseSchema}{The schema name of containing the vocabulary tables.}
}
\value{
Returns a tibble data frame.
}
\description{
convert a concept set expression object into a data frame object
}
