% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractConceptSetDetailsFromCohortExpression.R
\name{extractConceptSetDetailsFromCohortExpression}
\alias{extractConceptSetDetailsFromCohortExpression}
\title{Get concept set details from cohort definition}
\usage{
extractConceptSetDetailsFromCohortExpression(cohortExpression)
}
\arguments{
\item{cohortExpression}{A R-object (list) that represents cohort definition expression. This is derived from cohort expression json 
using RJSONIO::fromJSON(content = json, digits = 23). Note: it is important to use digits = 23, otherwise
numerical precision may be lost for large integer values like conceptId's in cohort definition. The cohort
expression JSON is commonly generated using OHDSI tools like Atlas or CapR.}
}
\description{
Get concept set details from cohort definition
}
