% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/ExtractConceptSetExpressionsFromCohortExpression.R
\name{extractConceptSetExpressionsFromCohortExpression}
\alias{extractConceptSetExpressionsFromCohortExpression}
\title{get concept set expressions from cohort expression}
\usage{
extractConceptSetExpressionsFromCohortExpression(cohortExpression)
}
\arguments{
\item{cohortExpression}{A R-object (list) that represents cohort definition expression. This is derived from cohort expression json 
using RJSONIO::fromJSON(content = json, digits = 23). Note: it is important to use digits = 23, otherwise
numerical precision may be lost for large integer values like conceptId's in cohort definition. The cohort
expression JSON is commonly generated using OHDSI tools like Atlas or CapR.}
}
\value{
Returns a tibble data frame with one row per concept set
}
\description{
given a cohort expression (R-list object, not JSON), this function parses the list and returns the concept set components
}
