% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractConceptSetDetailsFromCohortExpression.R
\name{extractConceptSetsSqlFromCohortSql}
\alias{extractConceptSetsSqlFromCohortSql}
\title{get concept set sql from cohort sql}
\usage{
extractConceptSetsSqlFromCohortSql(cohortSql)
}
\arguments{
\item{cohortSql}{sql of the cohort definition.}
}
\value{
Returns a tibble data frame with one row per concept set
}
\description{
given a cohort sql, this function parses the sql and return the concept set components
}
