% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindOrphanConceptsForConceptSetExpression.R
\name{findOrphanConceptsForConceptSetExpression}
\alias{findOrphanConceptsForConceptSetExpression}
\title{Find orphan concepts for a concept set expression.}
\usage{
findOrphanConceptsForConceptSetExpression(
  conceptSetExpression,
  vocabularyDatabaseSchema = "vocabulary",
  connection = NULL,
  connectionDetails = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema")
)
}
\arguments{
\item{vocabularyDatabaseSchema}{The schema name of containing the vocabulary tables.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}

\item{conceptIds}{An array of Concept ids.}
}
\value{
Returns a tibble data frame.
}
\description{
Find orphan concepts for a concept set expression.
}
