% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCohortSqlFromCohortExpressionUsingCirceR.R
\name{getCohortSqlFromCohortExpressionUsingCirceR}
\alias{getCohortSqlFromCohortExpressionUsingCirceR}
\title{get cohort sql from cohort expression}
\usage{
getCohortSqlFromCohortExpressionUsingCirceR(
  cohortExpression,
  generateStats = TRUE
)
}
\arguments{
\item{cohortExpression}{A R-object (list) that represents cohort definition expression. This is derived from cohort expression json 
using RJSONIO::fromJSON(content = json, digits = 23). Note: it is important to use digits = 23, otherwise
numerical precision may be lost for large integer values like conceptId's in cohort definition. The cohort
expression JSON is commonly generated using OHDSI tools like Atlas or CapR.}

\item{generateStats}{Do you want to include cohort generation stats tables in the SQL?}
}
\value{
Returns string (sql)
}
\description{
given a cohort expression R object, this function generates the cohort sql using circeR. The returned
SQL is a string that may be used with SqlRender.
}
