% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetConceptAncestor.R
\name{getConceptAncestor}
\alias{getConceptAncestor}
\title{given a list of conceptIds, get their ancestor and descendants}
\usage{
getConceptAncestor(
  conceptIds,
  connection,
  vocabularyDatabaseSchema = "vocabulary"
)
}
\arguments{
\item{conceptIds}{An array of Concept ids.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{vocabularyDatabaseSchema}{The schema name of containing the vocabulary tables.}
}
\description{
given a list of conceptIds, get their ancestor and descendants
}
