% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetConceptPrevalenceCountsForConceptIds.R
\name{getConceptPrevalenceCountsForConceptIds}
\alias{getConceptPrevalenceCountsForConceptIds}
\title{Get the count for an array of concept id(s) from concept prevalence table.}
\usage{
getConceptPrevalenceCountsForConceptIds(
  conceptIds,
  connection = NULL,
  connectionDetails = NULL,
  conceptPrevalenceSchema = "concept_prevalence"
)
}
\arguments{
\item{conceptIds}{An array of Concept ids.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{conceptPrevalenceSchema}{The schema containing the concept prevalence data.}
}
\description{
Get the count for an array of concept id(s) from concept prevalence table.
}
