% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetConceptSetDetailFromCohortDefinition.R
\name{getConceptSetDetailsFromCohortDefinition}
\alias{getConceptSetDetailsFromCohortDefinition}
\title{Get concept set details from cohort definition}
\usage{
getConceptSetDetailsFromCohortDefinition(cohortDefinitionExpression)
}
\arguments{
\item{cohortDefinitionExpression}{An R object (list) containing the expression of the OHDSI cohort definition.}
}
\description{
Get concept set details from cohort definition
}
