% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/GetConceptSetExpressionFromConceptSetExpressionDataFrame.R
\name{getConceptSetExpressionFromConceptSetExpressionDataFrame}
\alias{getConceptSetExpressionFromConceptSetExpressionDataFrame}
\title{Get concept set expression object from concept set expression data frame}
\usage{
getConceptSetExpressionFromConceptSetExpressionDataFrame(
  conceptSetExpressionDataFrame,
  selectAllDescendants = FALSE,
  purgeVocabularyDetails = FALSE
)
}
\arguments{
\item{conceptSetExpressionDataFrame}{Concept set expression in data frame format.}

\item{selectAllDescendants}{Do you want to over ride the concept set
expression by add select descendants for concept ids
in concept set expression.}

\item{purgeVocabularyDetails}{Do you want to purge the details of concepts in the
concept set expression.}
}
\description{
Get concept set expression object from concept set expression data frame
}
