% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetConceptSetRecommenderData.R
\name{getConceptSetRecommenderData}
\alias{getConceptSetRecommenderData}
\title{Get Concept Set Recommender Data}
\usage{
getConceptSetRecommenderData()
}
\value{
A data frame containing the concept set recommender data.
}
\description{
This function reads the concept recommender CSV file and returns the data as a data frame.
The CSV file is expected to be located in the "csv" folder with the file name
"concept_recommended_20240527.csv". After reading the data, the function converts the
column names from snake_case to camelCase using \code{SqlRender::snakeCaseToCamelCaseNames}.
}
