% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Private.R
\name{getDomainInformation}
\alias{getDomainInformation}
\title{Get domain information}
\usage{
getDomainInformation()
}
\value{
A list with two tibble data frame objects with domain information represented in wide and long format respectively.
}
\description{
Get domain information
}
