% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFuzzyMatchOfCodesToOmopConceptCode.R
\name{getFuzzyMatchOfCodesToOmopConceptCode}
\alias{getFuzzyMatchOfCodesToOmopConceptCode}
\title{Peform fuzzy match of codes.}
\usage{
getFuzzyMatchOfCodesToOmopConceptCode(
  sourceCodes,
  omopConcepts,
  removeSpecialCharacters = TRUE
)
}
\arguments{
\item{sourceCodes}{An array of codes to match to omop.}

\item{omopConcepts}{A data frame with atleast conceptId, conceptCode}

\item{removeSpecialCharacters}{During matching do you want to remove special characters.e.g. when
source codes have period removed, but they exist in omop vocabulary concept code.}
}
\value{
Returns a list of objects (to be described.)
}
\description{
Given two data frames, one with codes from external source (e.g. read from spreadsheet) vs
another data frame with OMOP vocabulary, perform fuzzy string match.
}
