% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/GetOptimizationRecommendationForConceptSetExpression.R
\name{getOptimizationRecommendationForConceptSetTable}
\alias{getOptimizationRecommendationForConceptSetTable}
\title{given a concept set tble, get optimization recommendation}
\usage{
getOptimizationRecommendationForConceptSetTable(
  conceptSetExpressionDataFrame,
  vocabularyDatabaseSchema = "vocabulary",
  connectionDetails = NULL,
  connection = NULL
)
}
\arguments{
\item{conceptSetExpressionDataFrame}{Concept set expression in data frame format.}

\item{vocabularyDatabaseSchema}{The schema name of containing the vocabulary tables.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}
}
\description{
given a concept set tble, get optimization recommendation
}
