% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOrphanConcepts.R
\name{getOrphanConcepts}
\alias{getOrphanConcepts}
\title{Get all the domain id(s) in the vocabulary schema.}
\usage{
getOrphanConcepts(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  tempEmulationSchema = NULL,
  conceptIds
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{vocabularyDatabaseSchema}{The schema name of containing the vocabulary tables.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}

\item{conceptIds}{an array of conceptIds that are used as reference to search for orphan concepts}
}
\value{
Returns a tibble data frame.
}
\description{
Get all the domain id(s) in the vocabulary schema.
}
