% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRecommendationForConceptSetExpression.R
\name{getRecommendationForConceptSetExpression}
\alias{getRecommendationForConceptSetExpression}
\title{Get recommended concepts for a concept set expression.}
\usage{
getRecommendationForConceptSetExpression(
  conceptSetExpression,
  vocabularyDatabaseSchema = "vocabulary",
  connection = NULL,
  connectionDetails = NULL,
  conceptPrevalenceSchema = "concept_prevalence",
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema")
)
}
\arguments{
\item{conceptSetExpression}{An R-object (list) with expression of the concept set.}

\item{vocabularyDatabaseSchema}{The schema name of containing the vocabulary tables.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{conceptPrevalenceSchema}{The schema name that has the concept prevalence table. The following
tables are expected to be present. recommender_set,
cp_master, recommended_blacklist.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}
}
\value{
Returns a tibble data frame.
}
\description{
Get recommended concepts for a concept set expression.
}
