% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSemanticConceptIds.R
\name{getSemanticConceptIds}
\alias{getSemanticConceptIds}
\title{Retrieve Concept IDs for Semantic Reasoning}
\usage{
getSemanticConceptIds(
  baseUrl,
  conceptSetDefinition,
  connection = NULL,
  connectionDetails = NULL,
  vocabularyDatabaseSchema,
  conceptRecommender
)
}
\arguments{
\item{baseUrl}{Character. The base URL for the web API.}

\item{conceptSetDefinition}{A concept set definition object used by ConceptSetDiagnostics.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{vocabularyDatabaseSchema}{The schema name of containing the vocabulary tables.}

\item{conceptRecommender}{Data frame. A data frame containing recommended concepts
used to filter additional concepts.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}
}
\value{
A list containing:
\describe{
  \item{resolvedMappedRecommended}{A data frame with the resolved, mapped, and recommended
  concepts along with associated concept names, ancestor, and synonym information.}
  \item{conceptIdDetails}{A data frame of detailed information for all collected concept IDs.}
  \item{recommendedConcepts}{A data frame of recommended concepts filtered based on the
  resolved concepts.}
}
}
\description{
This function retrieves concept IDs from a given concept set definition and processes
them for semantic reasoning. It performs several operations:
\itemize{
  \item Resolves the concept set definition.
  \item Retrieves mapped source concepts.
  \item Retrieves concept synonyms and ancestors.
  \item Filters and appends recommended concepts.
  \item Retrieves detailed information for all collected concept IDs.
  \item Merges the resolved, mapped, and recommended concepts with their ancestor and
        synonym details.
}
}
\details{
The function first establishes a database connection (if not provided) using the given
connection details. It then uses functions from the \code{ConceptSetDiagnostics} package to:
\enumerate{
  \item Resolve the provided concept set.
  \item Retrieve mapped source concepts.
  \item Retrieve synonyms and ancestors for each resolved concept.
}
The unique set of concept IDs is then used to filter recommended concepts and obtain
detailed concept information. Finally, the function merges the retrieved datasets and returns
a list of outputs for downstream semantic reasoning.
}
