% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/GetStandardMappingRecommendationsForNonStandard.R
\name{getStandardMappingRecommendationsForNonStandard}
\alias{getStandardMappingRecommendationsForNonStandard}
\title{Get standard mapping recommendations for non standard.}
\usage{
getStandardMappingRecommendationsForNonStandard(
  connectionDetails = NULL,
  connection = NULL,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  cdmDatabaseSchema = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  sourceVocabularyId = c("ICD10CM"),
  sourceCodes,
  removeSpecialCharacters = TRUE,
  includeDescendants = FALSE
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{vocabularyDatabaseSchema}{The schema name of containing the vocabulary tables.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}
}
\value{
Returns a list of objects (to be described.)
}
\description{
Get standard mapping recommendations for non standard.
}
