% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStringSearchConceptsUsingTsv.R
\name{getStringSearchConceptsUsingFullText}
\alias{getStringSearchConceptsUsingFullText}
\title{Get concepts that match a string search}
\usage{
getStringSearchConceptsUsingFullText(
  searchString,
  vocabularyDatabaseSchema = "vocabulary",
  connection = NULL,
  connectionDetails = NULL
)
}
\arguments{
\item{searchString}{A phrase (can be multiple words) to search for.}

\item{vocabularyDatabaseSchema}{The schema name of containing the vocabulary tables.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}
}
\description{
Get concepts that match a string search
}
