% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformConceptSetDiagnostics.R
\name{performConceptSetDiagnostics}
\alias{performConceptSetDiagnostics}
\title{given a search string (s) perform concept set diagnostics}
\usage{
performConceptSetDiagnostics(
  searchPhrases,
  connection = NULL,
  connectionDetails = NULL,
  conceptPrevalenceSchema = NULL,
  exportResults = FALSE,
  locationForResults = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  vocabularyDatabaseSchema = "vocabulary",
  vocabularyIdOfInterest = c("SNOMED", "HCPCS", "ICD10CM", "ICD10", "ICD9CM", "ICD9",
    "Read"),
  domainIdOfInterest = c("Condition", "Procedure", "Observation")
)
}
\arguments{
\item{searchPhrases}{An array of text-phrases within quotations that are seperated by commas.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{conceptPrevalenceSchema}{The schema name that has the concept prevalence table. The following
tables are expected to be present. recommender_set,
cp_master, recommended_blacklist.}

\item{exportResults}{Do you want to export results?}

\item{locationForResults}{If you want to export results, please provide disk drive location}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}

\item{vocabularyDatabaseSchema}{The schema name of containing the vocabulary tables.}

\item{vocabularyIdOfInterest}{A list of vocabulary ids to filter the results.}

\item{domainIdOfInterest}{A list of domain ids to filter the results.}
}
\description{
given an array of comma separated quoted search string (s), this function will perform a series 
of operations that provides a recommended concept set expression, along with
potentially more recommended and orphan concepts.
}
