% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformStringSearchForConcepts.R
\name{performStringSearchForConcepts}
\alias{performStringSearchForConcepts}
\title{Get concepts that match a string search}
\usage{
performStringSearchForConcepts(
  searchPhrases,
  vocabularyDatabaseSchema = "vocabulary",
  connection = NULL,
  connectionDetails = NULL,
  vocabularyIdOfInterest = c("SNOMED", "HCPCS", "ICD10CM", "ICD10", "ICD9CM", "ICD9",
    "Read"),
  domainIdOfInterest = c("Condition", "Procedure", "Observation"),
  retrieveInvalidConcepts = FALSE,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema")
)
}
\arguments{
\item{searchPhrases}{An array of phrases (each phrase may be multiple words,
but each phrase should be quoted and comma separated) to search for.
Phrases that have a character length of <4 are not searched.}

\item{vocabularyDatabaseSchema}{The schema name of containing the vocabulary tables.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{vocabularyIdOfInterest}{A list of vocabulary ids to filter the results.}

\item{domainIdOfInterest}{A list of domain ids to filter the results.}

\item{retrieveInvalidConcepts}{Do you want to retrieve invalid concepts. Default = FALSE}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}
}
\description{
Get concepts that match a string search
}
