% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetConceptIdDetails.R
\name{getConceptIdDetails}
\alias{getConceptIdDetails}
\title{get concept id details}
\usage{
getConceptIdDetails(
  conceptIds,
  connection = NULL,
  connectionDetails = NULL,
  vocabularyDatabaseSchema = "vocabulary",
  conceptPrevalenceTable = "concept_prevalence"
)
}
\arguments{
\item{conceptIds}{An array of Concept ids.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{vocabularyDatabaseSchema}{The schema name of containing the vocabulary tables.}

\item{conceptPrevalenceTable}{A reference table (can be schemaName.tableName) that
holds the concept prevalence data. The required fields are
concept_id, rc, drc, dbc, ddbc. In case of error,
cocneptPrevalence is silently ignored.}
}
\value{
Returns a tibble data frame.
}
\description{
given an array of conceptIds, get their details
}
