% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetConceptPrevalenceCounts.R
\name{getConceptPrevalenceCounts}
\alias{getConceptPrevalenceCounts}
\title{get concept id count}
\usage{
getConceptPrevalenceCounts(
  conceptIds,
  connection = NULL,
  connectionDetails = NULL,
  conceptPrevalenceTable,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema")
)
}
\arguments{
\item{conceptIds}{An array of Concept ids.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{conceptPrevalenceTable}{A reference table (can be schemaName.tableName) that
holds the concept prevalence data. The required fields are
concept_id, rc, drc, dbc, ddbc. In case of error,
cocneptPrevalence is silently ignored.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}
}
\value{
Returns a tibble data frame.
}
\description{
Get the count for an array of concept id(s) from concept prevalence table.
}
