% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/GetConceptSetExpressionDataFrameFromConceptSetExpression.R
\name{getConceptSetExpressionDataFrameFromConceptSetExpression}
\alias{getConceptSetExpressionDataFrameFromConceptSetExpression}
\title{convert a concept set expression object into a data frame object}
\usage{
getConceptSetExpressionDataFrameFromConceptSetExpression(
  conceptSetExpression,
  connection = NULL,
  connectionDetails = NULL,
  updateVocabularyFields = FALSE,
  vocabularyDatabaseSchema = "vocabulary"
)
}
\arguments{
\item{conceptSetExpression}{An R-object (list) with expression of the concept set.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{updateVocabularyFields}{Do you want to update the details of concepts from the vocabulary tables?}

\item{vocabularyDatabaseSchema}{The schema name of containing the vocabulary tables.}

\item{recordCount}{Do you want the function to return record count for the concept ids?}
}
\value{
Returns a tibble data frame.
}
\description{
convert a concept set expression object into a data frame object
}
