% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetConceptSetOccurrenceDate.R
\name{getConceptSetOccurrenceDate}
\alias{getConceptSetOccurrenceDate}
\title{Given a concept set expression, find the occurrence of the concept set}
\usage{
getConceptSetOccurrenceDate(
  connection,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  conceptIds,
  subset = c("all"),
  restrictToObservationPeriod = TRUE,
  tempEmulationSchema = NULL
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{vocabularyDatabaseSchema}{The schema name of containing the vocabulary tables.}

\item{conceptIds}{An array of concept ids}

\item{restrictToObservationPeriod}{(Default = TRUE) Do you want to restrict to Observation period? i.e
Cohort dates are restricted to observation period.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}
}
\description{
Given a concept set expression, find the occurrence of the concept set. The query will check
for the occurrence of any of the resolved conceptId's in the conceptSetExpression
in any domain tables. The domain tables are condition_occurrence, drug_exposure, procedure_occurrence,
visit_occurrence, observation.
Return objected is the name of the temp table
}
