% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDomain.R
\name{getDomain}
\alias{getDomain}
\title{Get all the domain id(s) in the vocabulary schema.}
\usage{
getDomain(connection, vocabularyDatabaseSchema = "vocabulary")
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{vocabularyDatabaseSchema}{The schema name of containing the vocabulary tables.}
}
\description{
Get all the domain id(s) in the vocabulary schema.
}
