% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRecommendationForConceptSetExpression.R
\name{getRecommendationForConceptSetExpression}
\alias{getRecommendationForConceptSetExpression}
\title{Get recommended concepts for a concept set expression.}
\usage{
getRecommendationForConceptSetExpression(
  conceptSetExpression,
  vocabularyDatabaseSchema = "vocabulary",
  vocabularyIdOfInterest = c("SNOMED", "HCPCS", "ICD10CM", "ICD10", "ICD9CM", "ICD9",
    "Read"),
  domainIdOfInterest = c("Condition", "Procedure", "Observation"),
  connection = NULL,
  connectionDetails = NULL,
  conceptPrevalenceSchema = "concept_prevalence"
)
}
\arguments{
\item{conceptSetExpression}{An R-object (list) with expression of the concept set.}

\item{vocabularyDatabaseSchema}{The schema name of containing the vocabulary tables.}

\item{vocabularyIdOfInterest}{A list of vocabulary id from OMOP to filter the results}

\item{domainIdOfInterest}{A list of domain id from OMOP to filter the results}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}
}
\description{
Get recommended concepts for a concept set expression.
}
