% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetStringSearchConcepts.R
\name{getStringSearchConcepts}
\alias{getStringSearchConcepts}
\title{Get concepts that match a string search}
\usage{
getStringSearchConcepts(
  searchString,
  fullTextSearch = FALSE,
  vocabularyDatabaseSchema = "vocabulary",
  connection
)
}
\arguments{
\item{searchString}{A phrase (can be multiple words) to search for.}

\item{fullTextSearch}{Do you want to do a full text search - this might take time.}

\item{vocabularyDatabaseSchema}{The schema name of containing the vocabulary tables.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}
}
\description{
Get concepts that match a string search
}
