% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetVocabulary.R
\name{getVocabulary}
\alias{getVocabulary}
\title{Get vocabulary id(s) in vocabulary tables in vocabulary schema.}
\usage{
getVocabulary(connection, vocabularyDatabaseSchema = "vocabulary")
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{vocabularyDatabaseSchema}{The schema name of containing the vocabulary tables.}
}
\description{
Get vocabulary id(s) in vocabulary tables in vocabulary schema.
}
