% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizeConceptSetExpression.R
\name{optimizeConceptSetExpression}
\alias{optimizeConceptSetExpression}
\title{given a concept set expression, get optimized concept set expression}
\usage{
optimizeConceptSetExpression(
  conceptSetExpression,
  vocabularyDatabaseSchema = "vocabulary",
  connection = NULL,
  connectionDetails = NULL
)
}
\arguments{
\item{conceptSetExpression}{An R-object (list) with expression of the concept set.}

\item{vocabularyDatabaseSchema}{The schema name of containing the vocabulary tables.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}
}
\description{
given a concept set expression, get optimized concept set expression
}
