% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResolveConceptSetsInCohortExpression.R
\name{resolveConceptSetsInCohortExpression}
\alias{resolveConceptSetsInCohortExpression}
\title{Given a cohort definition expression, get the resolved concepts for all concept sets}
\usage{
resolveConceptSetsInCohortExpression(
  cohortExpression,
  connection = NULL,
  connectionDetails = NULL,
  vocabularyDatabaseSchema = "vocabulary"
)
}
\arguments{
\item{cohortExpression}{A R-object (list) that represents cohort definition expression. This is derived from cohort expression json 
using RJSONIO::fromJSON(content = json, digits = 23). Note: it is important to use digits = 23, otherwise
numerical precision may be lost for large integer values like conceptId's in cohort definition. The cohort
expression JSON is commonly generated using OHDSI tools like Atlas or CapR.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{vocabularyDatabaseSchema}{The schema name of containing the vocabulary tables.}
}
\description{
Given a cohort definition expression, get the resolved concepts for all concept sets
}
